refactor(server): export helper
