+++ packages/core/server/src/index.ts
@@ -32,3 +32,4 @@ export {
 } from './plugin-manager/findPackageNames';
 
 export { runPluginStaticImports } from './run-plugin-static-imports';
+export { createContextVariablesScope } from './helper';
